assets/03-advanced/evaluation/requirements.txt
langfuse>=2.0.0
ragas>=0.1.0
openai>=1.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
